requirements.txt
opencv-python>=4.8.0
pillow>=10.0.0
requests>=2.31.0
openai>=1.3.0
anthropic>=0.8.0
google-generativeai>=0.3.0
click>=8.1.0
python-dotenv>=1.0.0
tqdm>=4.66.0
